backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pillow
moviepy
rembg
numpy
requests